refact: move getAllAppNames function from forge-fs to forge-process
